Guards and Logs
This change updates the code to use the guard construct.  This makes
the code much more readable and is more idiomatic Swift 2.  It also
modifies the logging statements to make the INFO level a bit less
verbose,  but moves some of the detailed information (specifically
locations of log files) into debug logging.